Google Ads Alert/requirements_google_ads.txt
google-ads>=23.0.0
google-cloud-bigquery>=3.0.0
google-cloud-secret-manager>=2.0.0
pandas-gbq>=0.18.0
pandas>=1.5.0
python-dotenv>=1.0.0
pytz>=2023.3